Move language logic to separate ecosystem
