+++ src/ecosystems/Global/index.js
@@ -2,6 +2,7 @@ export * from './WithApp';
 export * from './WithChildren';
 export * from './WithContent';
 export * from './WithErrorBoundary';
+export * from './WithLanguage';
 export * from './WithLoader';
 export * from './WithPubnub';
 export * from './WithQueryParams';
