+++ src/pages/App/NotFound.jsx
@@ -1,12 +1,14 @@
 import React from 'react';
-import { WithApp } from '../../ecosystems';
+import { WithApp, WithLanguage } from '../../ecosystems';
 import { DebugProps, AppNotFound } from '../../organisms';
 
 export const AppNotFoundPage = (props) => {
   return (
-    <WithApp {...props}>
-      <AppNotFound />
-      <DebugProps debug {...props} />
-    </WithApp>
+    <WithLanguage>
+      <WithApp {...props}>
+        <AppNotFound />
+        <DebugProps debug {...props} />
+      </WithApp>
+    </WithLanguage>
   );
 };
