+++ src/pages/CollectionItem/All.jsx
@@ -1,17 +1,25 @@
 import React from 'react';
-import { WithGithubAuth, WithGithubRepo, WithJekyll, WithApp } from '../../ecosystems';
+import {
+  WithApp,
+  WithGithubAuth,
+  WithGithubRepo,
+  WithJekyll,
+  WithLanguage,
+} from '../../ecosystems';
 import { CollectionList } from '../../organisms';
 
 export const CollectionItemAllPage = ({ children, ...props }) => {
   return (
     <WithGithubAuth {...props}>
-      <WithGithubRepo>
-        <WithJekyll>
-          <WithApp>
-            <CollectionList />
-          </WithApp>
-        </WithJekyll>
-      </WithGithubRepo>
+      <WithLanguage>
+        <WithGithubRepo>
+          <WithJekyll>
+            <WithApp>
+              <CollectionList />
+            </WithApp>
+          </WithJekyll>
+        </WithGithubRepo>
+      </WithLanguage>
     </WithGithubAuth>
   );
 };
