+++ src/pages/Pull/All.jsx
@@ -1,17 +1,25 @@
 import React from 'react';
-import { WithApp, WithGithubAuth, WithGithubRepo, WithJekyll } from '../../ecosystems';
+import {
+  WithApp,
+  WithGithubAuth,
+  WithGithubRepo,
+  WithJekyll,
+  WithLanguage
+} from '../../ecosystems';
 import { PullList } from '../../organisms';
 
 export const PullAllPage = ({ children, ...props }) => {
   return (
     <WithGithubAuth {...props}>
-      <WithGithubRepo>
-        <WithJekyll>
-          <WithApp>
-            <PullList />
-          </WithApp>
-        </WithJekyll>
-      </WithGithubRepo>
+      <WithLanguage>
+        <WithGithubRepo>
+          <WithJekyll>
+            <WithApp>
+              <PullList />
+            </WithApp>
+          </WithJekyll>
+        </WithGithubRepo>
+      </WithLanguage>
     </WithGithubAuth>
   );
 };
