+++ src/pages/Pull/One.jsx
@@ -1,15 +1,17 @@
 import React from 'react';
-import { WithApp, WithGithubAuth, WithGithubRepo } from '../../ecosystems';
+import { WithApp, WithGithubAuth, WithGithubRepo, WithLanguage } from '../../ecosystems';
 import { PullSingle } from '../../organisms';
 
 export const PullOnePage = ({ children, ...props }) => {
   return (
     <WithGithubAuth {...props}>
-      <WithGithubRepo>
-        <WithApp>
-          <PullSingle />
-        </WithApp>
-      </WithGithubRepo>
+      <WithLanguage>
+        <WithGithubRepo>
+          <WithApp>
+            <PullSingle />
+          </WithApp>
+        </WithGithubRepo>
+      </WithLanguage>
     </WithGithubAuth>
   );
 };
