+++ src/pages/Repo/One.jsx
@@ -1,17 +1,25 @@
 import React from 'react';
-import { WithApp, WithGithubAuth, WithGithubRepo, WithJekyll } from '../../ecosystems';
+import {
+  WithApp,
+  WithGithubAuth,
+  WithGithubRepo,
+  WithJekyll,
+  WithLanguage
+} from '../../ecosystems';
 import { RepoSingle } from '../../organisms';
 
 export const RepoOnePage = ({ children, ...props }) => {
   return (
     <WithGithubAuth {...props}>
-      <WithGithubRepo>
-        <WithJekyll>
-          <WithApp>
-            <RepoSingle />
-          </WithApp>
-        </WithJekyll>
-      </WithGithubRepo>
+      <WithLanguage>
+        <WithGithubRepo>
+          <WithJekyll>
+            <WithApp>
+              <RepoSingle />
+            </WithApp>
+          </WithJekyll>
+        </WithGithubRepo>
+      </WithLanguage>
     </WithGithubAuth>
   );
 };
