+++ src/pages/User/All.jsx
@@ -1,19 +1,28 @@
 import React from 'react';
-import { WithApp, WithGithubAuth, WithGithubRepo, WithJekyll, WithGithubRepoUsers } from '../../ecosystems';
+import {
+  WithApp,
+  WithGithubAuth,
+  WithGithubRepo,
+  WithGithubRepoUsers,
+  WithJekyll,
+  WithLanguage
+} from '../../ecosystems';
 import { UserList } from '../../organisms';
 
 export const UserAllPage = ({ children, ...props }) => {
   return (
     <WithGithubAuth {...props}>
-      <WithGithubRepo>
-        <WithJekyll>
-          <WithGithubRepoUsers>
-            <WithApp>
-              <UserList />
-            </WithApp>
-          </WithGithubRepoUsers>
-        </WithJekyll>
-      </WithGithubRepo>
+      <WithLanguage>
+        <WithGithubRepo>
+          <WithJekyll>
+            <WithGithubRepoUsers>
+              <WithApp>
+                <UserList />
+              </WithApp>
+            </WithGithubRepoUsers>
+          </WithJekyll>
+        </WithGithubRepo>
+      </WithLanguage>
     </WithGithubAuth>
   );
 };
